# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/roles": {
      "get": {
        "description": "Este endpoint lista los roles activos con soporte para paginación",
        "operationId": "RolesController_getAllRoles",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Obtener todos los roles con paginación",
        "tags": [
          "Roles"
        ]
      },
      "post": {
        "description": "Este endpoint sirve para crear nuevos roles",
        "operationId": "RolesController_createRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Para crear un nuevo rol",
        "tags": [
          "Roles"
        ]
      }
    },
    "/roles/{idRol}": {
      "get": {
        "description": "Este endpoint sirve para obtener un rol en específico",
        "operationId": "RolesController_getRole",
        "parameters": [
          {
            "name": "idRol",
            "required": true,
            "in": "path",
            "description": "Id del rol a obtener",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Obtener un rol por su id",
        "tags": [
          "Roles"
        ]
      },
      "patch": {
        "description": "Este endpoint nos sirve para poder actualizar un rol",
        "operationId": "RolesController_updateRole",
        "parameters": [
          {
            "name": "idRol",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Actualizar un rol existente",
        "tags": [
          "Roles"
        ]
      }
    },
    "/roles/{idRol}/delete": {
      "patch": {
        "description": "Este endpoint nos sirve para desactivar un rol",
        "operationId": "RolesController_deleteRole",
        "parameters": [
          {
            "name": "idRol",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Eliminar un rol a elección",
        "tags": [
          "Roles"
        ]
      }
    },
...
    "/stock": {
      "get": {
        "operationId": "StockController_getAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de stock obtenida correctamente"
          }
        },
        "summary": "Obtener todos los registros de stock activos con paginación",
        "tags": [
          "stock"
        ]
      },
      "post": {
        "operationId": "StockController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStockDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Stock creado exitosamente"
          },
          "400": {
            "description": "Datos inválidos para creación"
          }
        },
        "summary": "Crear un nuevo registro de stock",
        "tags": [
          "stock"
        ]
      }
    },
    "/stock/{id}": {
      "get": {
        "operationId": "StockController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del registro de stock",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stock encontrado exitosamente"
          },
          "404": {
            "description": "Registro no encontrado"
          }
        },
        "summary": "Obtener un registro de stock por ID",
        "tags": [
          "stock"
        ]
      },
      "patch": {
        "operationId": "StockController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStockDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stock actualizado exitosamente"
          },
          "400": {
            "description": "Datos inválidos para actualización"
          },
          "404": {
            "description": "Registro no encontrado para actualizar"
          }
        },
        "summary": "Actualizar un registro de stock existente",
        "tags": [
          "stock"
        ]
      }
    },
    "/stock/{id}/delete": {
      "patch": {
        "operationId": "StockController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stock desactivado correctamente"
          },
          "404": {
            "description": "Registro no encontrado o ya desactivado"
          }
        },
        "summary": "Desactivar un registro de stock (borrado lógico)",
        "tags": [
          "stock"
        ]
      }
    },
...
  },
  "info": {
    "title": "Control de Endpoints",
    "description": "Documentación de la API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "nombreRol": {
            "type": "string",
            "description": "Nombre del rol",
            "example": "Administrador"
          }
        },
        "required": [
          "nombreRol"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "nombreRol": {
            "type": "string",
            "description": "Nombre del rol a actualizar",
            "example": "Administrador corregido"
          },
          "estado": {
            "type": "number",
            "description": "Estado del rol (1 = activo, 0 = inactivo)",
            "example": 1
          }
        }
      },
...
      "CreateStockDto": {
        "type": "object",
        "properties": {
          "idProducto": {
            "type": "number",
            "description": "id del producto relacionado (ej. iPhone 13 Pro)",
            "example": 1
          },
          "idLote": {
            "type": "number",
            "description": "id del lote relacionado (ej. lote de julio 2025)",
            "example": 3
          },
          "cantidadStock": {
            "type": "number",
            "description": "Cantidad de stock disponible para este producto en el lote",
            "example": 25
          }
        },
        "required": [
          "idProducto",
          "idLote",
          "cantidadStock"
        ]
      },
      "UpdateStockDto": {
        "type": "object",
        "properties": {
          "idProducto": {
            "type": "number",
            "description": "ID del producto relacionado",
            "example": 1
          },
          "idLote": {
            "type": "number",
            "description": "ID del lote relacionado",
            "example": 1
          },
          "cantidadStock": {
            "type": "number",
            "description": "Nueva cantidad de stock",
            "example": 50
          },
          "estado": {
            "type": "boolean",
            "description": "Estado del registro de stock",
            "example": true
          }
        }
      },
...
    }
  }
}